refactor: TxClient
