+++ .changeset/stupid-dodos-fold.md
@@ -1,0 +1,5 @@
+---
+'@bnb-chain/greenfield-js-sdk': patch
+---
+
+refactor: basic.multi -> txClient.multi
