+++ packages/chain-sdk/src/api/account.ts
@@ -1,3 +1,4 @@
+import { TxClient } from '@/clients/txClient';
 import { MsgMultiSendSDKTypeEIP712 } from '@/messages/bank/MsgMultiSend';
 import { MsgSendSDKTypeEIP712 } from '@/messages/bank/MsgSend';
 import { MsgCreatePaymentAccountSDKTypeEIP712 } from '@/messages/greenfield/payment/MsgCreatePaymentAccount';
@@ -24,7 +25,6 @@ import {
   MsgSendTypeUrl,
   TxResponse,
 } from '..';
-import { Basic } from './basic';
 import { RpcQueryClient } from '../clients/queryclient';
 
 export interface IAccount {
@@ -67,12 +67,12 @@ export interface IAccount {
 
 @singleton()
 export class Account implements IAccount {
-  constructor(@inject(delay(() => Basic)) private basic: Basic) {}
+  constructor(@inject(delay(() => TxClient)) private txClient: TxClient) {}
 
   private queryClient = container.resolve(RpcQueryClient);
 
   public async multiTransfer(address: string, msg: MsgMultiSend) {
-    return await this.basic.tx(
+    return await this.txClient.tx(
       MsgMultiSendTypeUrl,
       address,
       MsgMultiSendSDKTypeEIP712,
@@ -82,7 +82,7 @@ export class Account implements IAccount {
   }
 
   public async createPaymentAccount(msg: MsgCreatePaymentAccount) {
-    return await this.basic.tx(
+    return await this.txClient.tx(
       MsgCreatePaymentAccountTypeUrl,
       msg.creator,
       MsgCreatePaymentAccountSDKTypeEIP712,
@@ -131,7 +131,7 @@ export class Account implements IAccount {
   }
 
   public async transfer(msg: MsgSend) {
-    return await this.basic.tx(
+    return await this.txClient.tx(
       MsgSendTypeUrl,
       msg.fromAddress,
       MsgSendSDKTypeEIP712,
