+++ packages/chain-sdk/src/api/bucket.ts
@@ -21,6 +21,7 @@ import {
   getQueryBucketReadQuotaMetaInfo,
   parseReadQuotaResponse,
 } from '@/clients/spclient/spApis/queryBucketReadQuota';
+import { TxClient } from '@/clients/txClient';
 import { METHOD_GET, NORMAL_ERROR_CODE } from '@/constants/http';
 import { MsgCreateBucketSDKTypeEIP712 } from '@/messages/greenfield/storage/MsgCreateBucket';
 import { MsgDeleteBucketSDKTypeEIP712 } from '@/messages/greenfield/storage/MsgDeleteBucket';
@@ -86,7 +87,6 @@ import {
 } from '..';
 import { RpcQueryClient } from '../clients/queryclient';
 import { AuthType, SpClient } from '../clients/spclient/spClient';
-import { Basic } from './basic';
 import { Sp } from './sp';
 import { Storage } from './storage';
 
@@ -173,7 +173,7 @@ export interface IBucket {
 @singleton()
 export class Bucket implements IBucket {
   constructor(
-    @inject(delay(() => Basic)) private basic: Basic,
+    @inject(delay(() => TxClient)) private txClient: TxClient,
     @inject(delay(() => Sp)) private sp: Sp,
     @inject(delay(() => Storage)) private storage: Storage,
   ) {}
@@ -252,7 +252,7 @@ export class Bucket implements IBucket {
   }
 
   private async createBucketTx(msg: MsgCreateBucket, signedMsg: CreateBucketApprovalResponse) {
-    return await this.basic.tx(
+    return await this.txClient.tx(
       MsgCreateBucketTypeUrl,
       msg.creator,
       MsgCreateBucketSDKTypeEIP712,
@@ -294,7 +294,7 @@ export class Bucket implements IBucket {
   }
 
   public async deleteBucket(msg: MsgDeleteBucket) {
-    return await this.basic.tx(
+    return await this.txClient.tx(
       MsgDeleteBucketTypeUrl,
       msg.operator,
       MsgDeleteBucketSDKTypeEIP712,
@@ -451,7 +451,7 @@ export class Bucket implements IBucket {
       }),
     };
 
-    return await this.basic.tx(
+    return await this.txClient.tx(
       MsgUpdateBucketInfoTypeUrl,
       msg.operator,
       MsgUpdateBucketInfoSDKTypeEIP712,
@@ -569,7 +569,7 @@ export class Bucket implements IBucket {
   }
 
   private async migrateBucketTx(msg: MsgMigrateBucket, signedMsg: MigrateBucketApprovalResponse) {
-    return await this.basic.tx(
+    return await this.txClient.tx(
       MsgMigrateBucketTypeUrl,
       msg.operator,
       MsgMigrateBucketSDKTypeEIP712,
